Move the definition of Singleton out of the config module, as it's not related to config.
